Lang: Add stage 2 & 3 for match-statements with list-shape-support
In an old version in another repo, I used labels and goto heavily for
this. As I want to target JS at some point and JS doesn't have those
things, we instead just copy-paste the code where we would have jumped
to in any reasonable implementation. This leads to massive duplicate
code being generated, but it works.