Lang: WIP Generate SamalAST.hx from template file
This will make it easier to add nodes in the future as you need to type
way less code. In addition to that, I can directly integrate clone
methods, making the cloner-library superfluous which should also fix a
bug in the python target. As there are no more abstract classes (for
now) this should fix a bug in the C++ target as well.